refactor: make utility function out of repeated media lines
